+++ fresh.gen.ts
@@ -4,6 +4,10 @@
 
 import * as $_404 from "./routes/_404.tsx";
 import * as $_app from "./routes/_app.tsx";
+import * as $api_chips_add from "./routes/api/chips/add.ts";
+import * as $api_chips_edit from "./routes/api/chips/edit.ts";
+import * as $api_chips_get from "./routes/api/chips/get.ts";
+import * as $api_chips_list from "./routes/api/chips/list.ts";
 import * as $button_index from "./routes/button/index.tsx";
 import * as $chip_index from "./routes/chip/index.tsx";
 import * as $demo from "./routes/demo.tsx";
@@ -15,13 +19,17 @@ import * as $layout_index from "./routes/layout/index.tsx";
 import * as $markdown_index from "./routes/markdown/index.tsx";
 import * as $panel_panel from "./routes/panel/panel.tsx";
 import * as $text_text from "./routes/text/text.tsx";
-import * as $lunchbox from "./islands/lunchbox.ts";
+import * as $AddChip from "./islands/AddChip.tsx";
 import { type Manifest } from "$fresh/server.ts";
 
 const manifest = {
   routes: {
     "./routes/_404.tsx": $_404,
     "./routes/_app.tsx": $_app,
+    "./routes/api/chips/add.ts": $api_chips_add,
+    "./routes/api/chips/edit.ts": $api_chips_edit,
+    "./routes/api/chips/get.ts": $api_chips_get,
+    "./routes/api/chips/list.ts": $api_chips_list,
     "./routes/button/index.tsx": $button_index,
     "./routes/chip/index.tsx": $chip_index,
     "./routes/demo.tsx": $demo,
@@ -35,7 +43,7 @@ const manifest = {
     "./routes/text/text.tsx": $text_text,
   },
   islands: {
-    "./islands/lunchbox.ts": $lunchbox,
+    "./islands/AddChip.tsx": $AddChip,
   },
   baseUrl: import.meta.url,
 } satisfies Manifest;