+++ islands/lunchbox.ts
@@ -1,1 +1,0 @@
-export * from 'lunchbox/islands.ts';