+++ routes/chip/index.tsx
@@ -1,35 +1,26 @@
-import {
-  Card,
-  Chip,
-  Header,
-  Main,
-  Separator,
-  Text,
-} from 'lunchbox/components.ts';
-import { ChipSelect } from 'lunchbox/islands.ts';
+import Card from "lunchbox/components/Card/index.tsx";
+import Chip from "lunchbox/components/Chip/index.tsx";
+import Header from "lunchbox/components/Header/index.tsx";
+import Main from "lunchbox/components/Main/index.tsx";
+import Separator from "lunchbox/components/Separator/index.tsx";
+import Text from "lunchbox/components/Text/index.tsx";
+import Code from "lunchbox/components/Code/index.tsx";
+import AddChip from "../../islands/AddChip.tsx";
+import { dbChip, type modChip } from "@/db.ts";
 
 export default function ChipTest() {
   return (
     <>
-      <Header layout_type='focus'>
-        <Text type='title'>Chip tests</Text>
+      <Header layout_type="focus">
+        <Text type="title">Chip tests</Text>
       </Header>
-      <Main layout_type='focus'>
-        <Card>
-          <Text type='subheading' noMargins>Chip</Text>
-          <div class='flex gap-4'>
-            <Chip content='Test Chip' />
-            <Chip content='Selected Chip' selected />
-          </div>
-          <Separator />
-          <Text type='subheading' noMargins>ChiplistSelect</Text>
-          <ChipSelect
-            initialChips={[
-              { content: 'test chip' },
-              { content: 'test 2 chip' },
-            ]}
-          />
-        </Card>
+      <Main layout_type="focus">
+        <>
+          <Text type="subheading" noMargins>
+            Adding chips to <Code>DenoKv</Code>
+          </Text>
+          <AddChip<modChip> initialValues={[]} />
+        </>
       </Main>
     </>
   );