+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## v0.0.3
+
+- Added a module for managing the database. `/db.ts`
+- Added `--unstable-kv` to `deno task start`.
+- Added the `<AppChip/>` island.
+- Removed the lunchbox island export module.
+- Added CRUD endpoints for the chip model.
+- Refactored the chip test site.
+
 ## v0.0.2
 
 - Added necessary changes for local development.